Bug 2062075 - Move the pixelSizeFor helpers into android-components. r=android-reviewers,twhite
These wrap a call that every project makes, but they were written in Fenix, so nobody else could reach
them. Move them next to the other extensions on the same receivers and leave the behaviour alone; the
lint check that recommends them stays restricted to Fenix for now.

Differential Revision: https://phabricator.services.mozilla.com/D317360
