+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/components/toolbar/ToolbarsIntegration.kt
@@ -12,8 +12,8 @@ import mozilla.components.concept.engine.EngineView
 import mozilla.components.feature.pwa.feature.WebAppHideToolbarFeature
 import mozilla.components.feature.session.FullScreenFeature
 import mozilla.components.support.base.feature.LifecycleAwareFeature
+import mozilla.components.support.ktx.android.view.pixelSizeFor
 import mozilla.components.support.utils.ext.isKeyboardVisible
-import org.mozilla.fenix.ext.pixelSizeFor
 import org.mozilla.fenix.utils.Settings
 import kotlin.math.roundToInt
 import mozilla.components.compose.browser.toolbar.R as toolbarR
