+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/ext/Context.kt
@@ -8,14 +8,12 @@ import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.content.res.Configuration
-import android.content.res.Resources
 import android.net.Uri
 import android.provider.Settings
 import android.view.ContextThemeWrapper
 import android.view.View
 import android.view.ViewGroup
 import android.view.accessibility.AccessibilityManager
-import androidx.annotation.DimenRes
 import androidx.annotation.StringRes
 import mozilla.components.compose.base.theme.layout.AcornWindowSize
 import mozilla.components.support.base.log.logger.Logger
@@ -216,17 +214,3 @@ fun Context.recordEventInNimbus(eventId: String) = components.nimbus.events.reco
  */
 fun Context.isToolbarAtBottom() =
     components.settings.toolbarPosition == ToolbarPosition.BOTTOM
-
-/**
- * Returns the pixel size for the given dimension resource ID.
- *
- * This is a wrapper around [Resources.getDimensionPixelSize], reducing verbosity when accessing
- * dimension values from a [Context].
- *
- * @param resId Resource ID of the dimension.
- * @return The pixel size corresponding to the given dimension resource.
- */
-@Suppress("Resources.GetDimensionPixelSizeInsteadOfPixelSizeFor")
-fun Context.pixelSizeFor(
-    @DimenRes resId: Int,
-) = resources.getDimensionPixelSize(resId)
